chore: remove unused vi import
