+++ web/src/vocab/TextInput.tsx
@@ -1,4 +1,4 @@
-import { ChangeEvent } from 'react'
+import type { ChangeEvent } from 'react'
 import useVocabAnswer from './useVocabAnswer'
 
 const TextInput = () => {
